HdrHistogram explicit Java module.
Motivation:

Support explicit Java modules in HdrHistogram while continue providing a Java 8 baseline. This aims to bring modularity to the Java ecosystem of libraries.

Changes:

Use the io.github.dmlloyd.module-info:module-info plugin that produces a module-info class file that can be used when HdrHistogram is used on the module path.

Tests have also been added, and is only executed with a Java 11 profile. These tests ensures that HdrHistogram can be used as a module on the module path and that it is actually in an explicit module.